% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqmr.R
\docType{data}
\name{corpus_bara_turn}
\alias{corpus_bara_turn}
\title{UK Abortion Debate in Speaker Contribution Form}
\format{A quanteda corpus object with docvars
\describe{
  \item{speaker}{Speaker of each paragraph}
  \item{vote}{Which way this speaker voted after the debate:
              "yes", "no", or "abs" (abstained)}
}}
\source{
\url{https://api.parliament.uk/historic-hansard/commons/1966/jul/22/medical-termination-of-pregnancy-bill}
}
\usage{
corpus_bara_turn
}
\description{
Second reading of David Steele's 1966 bill to relax UK Abortion law in 1966,
scraped from Hansard and aggregated by speaker 'turn'.
}
\keyword{datasets}
