% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqmr.R
\docType{data}
\name{corpus_uk_manif}
\alias{corpus_uk_manif}
\title{Post War UK Party Platforms}
\format{A quanteda corpus object with docvars
\describe{
  \item{country}{Always 'UK'}
  \item{scope}{Always 'natl' (national)}
  \item{year}{Year of election}
  \item{lang}{Always 'en' (English)}
  \item{party}{Party name abbreviation. See Details for the main parties}
}}
\source{
Ken Benoit and the original 'quantedaData' package
}
\usage{
corpus_uk_manif
}
\description{
UK Party platforms for general election from 1945 to 2015.
From 1997 includes minor and regional parties.  Main parties are
"Con" (Conservatives), "Lab" (Labour), and "Lib" (Liberals),
"LibSDP" (Liberals / SDP Alliance),
"LD" (Liberal Democrats).  "Coalition" denotes the Conservative-Liberal
Democrat coalition government.
}
\keyword{datasets}
