% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqmr.R
\docType{data}
\name{df_pba_debate}
\alias{df_pba_debate}
\title{Partial-Birth Abortion Act Senate Debate in Paragraph Form}
\format{A data frame object with variables
\describe{
  \item{speaker}{Speaker's family name, capitalized except for DeWine}
  \item{text}{What was spoken}
  \item{quote}{Whether this part of the record is the speaker's own words
               or a quote, either of the bill itself or a letter from
               another person or organization}
  \item{party}{Party membership of the speaker. "D" (Democrat) or
               "R" (Republican).  For the 'presiding officer', no party
               is recorded although they will have one.}
}}
\source{
Congressional Record:
        \url{https://www.congress.gov/crec/2003/10/21/modified/CREC-2003-10-21-pt1-PgS12914-2.htm},
        \url{https://www.congress.gov/crec/2003/10/21/modified/CREC-2003-10-21-pt1-PgS12927-2.htm}
}
\usage{
df_pba_debate
}
\description{
The final US Senate debate for the 2003 Act banning 'partial-birth' abortion,
scraped from the congressional record and organized by paragraph (or
a near approximation).
}
\details{
This is the kind of data frame you might make a quanteda corpus out of.
}
\keyword{datasets}
