% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqmr.R
\docType{data}
\name{df_pba_debate_by_speaker}
\alias{df_pba_debate_by_speaker}
\title{Partial-Birth Abortion Act Senate Debate in Speaker Contribution Form}
\format{A data frame object with variables
\describe{
  \item{speaker}{Speaker's family name, capitalized except for DeWine}
  \item{contributions}{All this speaker said in the course of debate}
  \item{party}{Party membership of the speaker. "D" (Democrat) or
               "R" (Republican).}
}}
\source{
Congressional Record:
        \url{https://www.congress.gov/crec/2003/10/21/modified/CREC-2003-10-21-pt1-PgS12914-2.htm},
        \url{https://www.congress.gov/crec/2003/10/21/modified/CREC-2003-10-21-pt1-PgS12927-2.htm}
}
\usage{
df_pba_debate_by_speaker
}
\description{
The final US Senate debate for the 2003 Act banning 'partial-birth' abortion,
scraped from the Congressional Record and organized by paragraph (or
a near approximation).
}
\details{
This is the kind of data frame you might make a quanteda corpus out of.
It was constructed by filtering and aggregating \code{\link{df_pba_debate}}.
Contributions from the sequence of 'presiding officers' has been removed.
}
\keyword{datasets}
