% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqmr.R
\docType{data}
\name{dict_baraetal}
\alias{dict_baraetal}
\title{Bara et al. Dictionary}
\format{A quanteda dictionary}
\source{
Judith Bara
}
\usage{
dict_baraetal
}
\description{
The content analysis dictionary used in J. Bara, A. Weale, and A. Bicquelet
(2007) 'Analysing Parliamentary Debate with Computer Assistance'
Swiss Political Science Review 13:4
\url{https://doi.org/10.1002/j.1662-6370.2007.tb00090.x}.
}
\details{
This dictionary can be used to analyse \code{\link{corpus_bara_speaker}}
}
\keyword{datasets}
