% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqmr.R
\docType{data}
\name{dict_lavergarry}
\alias{dict_lavergarry}
\title{Laver and Garry Dictionary}
\format{A quanteda dictionary}
\source{
Michael Laver
}
\usage{
dict_lavergarry
}
\description{
The content analysis dictionary used in M. Laver and J. Garry (2000)
'Estimating Policy Poisitions from Political Text'
American Journal of Political Science 44:3
\url{https://doi.org/10.2307/2669268}.
}
\details{
This dictionary was constructed from the platforms of the three main
parties contesting the 1992 UK election.  These manifestos can be found
in \code{\link{corpus_uk_manif}}.
}
\keyword{datasets}
