% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqmr.R
\docType{package}
\name{iqmr}
\alias{iqmr}
\alias{iqmr-package}
\title{iqmr: Materials for the IQMR summer school}
\description{
The iqmr package provides materials for the IQMR summer school at the
Maxwell School at the University of Syracuse.  It is mostly a
repository of data but also stores labs and slides as vignettes.
To access lecture materials by name, use \code{\link{show_lecture}}.
To access lab instructions by name, use \code{\link{show_lab}}.
}
\section{Naming convention}{

Data frames are prefixed \code{df_}, quanteda corpus objects are prefixed
\code{corpus_}, and quanteda dictionaries are prefixed \code{dict_}.
}

