% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqmr.R
\name{show_lecture}
\alias{show_lecture}
\title{Show a set of slides}
\usage{
show_lecture(which = c(1, 2, 3))
}
\arguments{
\item{which}{Which session's lecture slides to show: Currently 1, 2, or 3}
}
\value{
None
}
\description{
Show a set of slides
}
